# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCategory } from "./productInputCategory";

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  price: number;
  category: ProductInputCategory;
  /** @minLength 1 */
  image: string;
  description: string;
  sizes: string[];
  /**
   * @minimum 0
   * @maximum 100
   */
  popularity: number;
}