# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContent {
  heroHeadline: string;
  heroSubheading: string;
  heroCtaPrimary: string;
  heroCtaSecondary: string;
  whatsappNumber: string;
  contactEmail: string;
}