# lib/api-zod/src/generated/types/siteContentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContentInput {
  /** @minLength 1 */
  heroHeadline: string;
  /** @minLength 1 */
  heroSubheading: string;
  /** @minLength 1 */
  heroCtaPrimary: string;
  /** @minLength 1 */
  heroCtaSecondary: string;
  /** @minLength 1 */
  whatsappNumber: string;
  /** @minLength 1 */
  contactEmail: string;
}